Refactor Pulse Level Abstraction Code
* Files associated with the Pulse Level Abstraction Code are now stored in the directory pulse_level_abstraction on the highest level of the repo
* Clean out the former directory ./template_configurations/qiskit/pulse_level
