+++ pulse_level_abstraction/training/q_env_pulse_config.yml
@@ -3,7 +3,7 @@ SERVICE: # Relevant only when using the Qiskit runtime service
   INSTANCE: "ibm-q-nus/default/default"
 
 RUNTIME_OPTIONS: # Relevant only when using the Qiskit runtime service
-  optimization_level: 3
+  optimization_level: 0
   resilience_level: null
   max_execution_time: null
   execution:
@@ -30,19 +30,13 @@ RUNTIME_OPTIONS: # Relevant only when using the Qiskit runtime service
   environment:
     log_level: "WARNING"
     job_tags: null
-  simulator:
-    seed_simulator: 100
-    coupling_map: null
-    basis_gates: null
 
 BACKEND: # Backend configuration (If all set to null, the user needs to specify its own backend in q_env_config.py's get_backend() function)
-  REAL_BACKEND: null # True: real or False: fake Aer backend
-  NAME: null # Name of the backend
-  REAL_BACKEND: null # True: real or False: fake Aer backend
-  NAME: null # Name of the backend
+  REAL_BACKEND: True # True: real or False: fake Aer backend
+  NAME: "ibm_nazca" # Name of the backend
   DYNAMICS: # Use a DynamicsBackend (if fields above are not null, build a DynamicsBackend.from_backend() with the specified backend)
-    USE_DYNAMICS: True # Whether to use a DynamicsBackend
-    PHYSICAL_QUBITS: [ 0 ] # qubit indices of the device to be used for modelling the system
+    USE_DYNAMICS: null # Whether to use a DynamicsBackend
+    PHYSICAL_QUBITS: [ 97, 98 ] # qubit indices of the device to be used for modelling the system
     SOLVER_OPTIONS:
       method: "jax_odeint"
       atol: 1e-5
@@ -52,25 +46,27 @@ BACKEND: # Backend configuration (If all set to null, the user needs to specify 
 
 
 TARGET: # Target Gate configuration
-  # GATE: "X" # Gate to be implemented (if null, the target state should be specified)
-  STATE: "1" # Target state (if GATE is null)
-  PHYSICAL_QUBITS: [ 0 ] # qubit indices for the target gate (qubits for which the gate should be calibrated)
+  GATE: "ECR" # Gate to be implemented (if null, the target state should be specified)
+  #  STATE: "1" # Target state (if GATE is null)
+  PHYSICAL_QUBITS: [ 97, 98 ] # qubit indices for the target gate (qubits for which the gate should be calibrated)
 
 ENV: # Environment configuration
-  SAMPLING_PAULIS: 50 # Number of Pauli strings to sample
-  N_SHOTS: 10
-  N_REPS: 1 # Number of repetitions for the fidelity benchmarking
-  N_REPS: 1 # Number of repetitions for the fidelity benchmarking
-  C_FACTOR: 0.5 # Cost factor for the reward function
-  BATCH_SIZE: 256 # Number of actions to evaluate per policy iteration
+  EXECUTION:
+    SAMPLING_PAULIS: 80 # Number of Pauli strings to sample
+    N_SHOTS: 1 # Number of shots for each Pauli
+    N_REPS: 1 # Number of repetitions for the fidelity benchmarking
+    C_FACTOR: 1. # Cost factor for the reward function
+    BATCH_SIZE: 256 # Number of actions to evaluate per policy iteration
+    SEED: 100
   ACTION_SPACE:
-    LOW: [ -1., -3.14 ]
-    HIGH: [ 1., 3.14 ]
-  BENCHMARK_CYCLE: 10 # Number of steps between two fidelity benchmarks
-  SEED: 100
-  CHECK_ON_EXP: False # Whether to perform fidelity benchmarking with tomographic experiments or just using simulation
-  TRAINING_WITH_CAL: False
-  CHANNEL_ESTIMATOR: False
-  FIDELITY_ACCESS: True
-  FIDELITY_ACCESS: True
+    LOW: [ -1., -3.14, -1., -3.14 ]
+    HIGH: [ 1., 3.14, 1., 3.14 ]
+  REWARD:
+    REWARD_METHOD: "state" # Choose between "fidelity", "state", "channel", "xeb", "cafe"
+  BENCHMARKING:
+    BENCHMARK_CYCLE: 20 # Number of steps between two fidelity benchmarks
+    BENCHMARK_BATCH_SIZE: 20 # Number of actions to evaluate per fidelity benchmark
+    CHECK_ON_EXP: True # Whether to perform fidelity benchmarking with tomographic experiments or just using simulation
+    TOMOGRAPHY_ANALYSIS: "default" # Analysis method for tomography experiment
+  TRAINING_WITH_CAL: True
